+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (Unreleased)
 
+-   Support helm v3 `include-crds` argument. (https://github.com/pulumi/pulumi-kubernetes/pull/1102)
+
 ## 2.1.1 (May 8, 2020)
 
 -   Python and .NET packages failed to publish for 2.1.0, so bumping release version.